Extend and improve the testsuite, more than doubling the number of tests.
Instead of just checking the default parameter for each backend,
create multiple sub-cases for all the tests. The Standalone case will
check the behaviour of the backend without an actual instance open,
while the others will test that the backends work properly with all
the possible variation in valid parameters they are given.